post merge fixes + proper timeout calculation
tnow = static_cast<uint32_t>(millis());
if (tnow - time > timeout) {

deals with buffer overflow and forces all to be uint32

+++ arduino/common/lib/CommsControl/CommsControl.h
@@ -57,7 +57,7 @@ private:
 
     uint32_t _baudrate;
 
-    uint64_t _last_trans_time;
+    uint32_t _last_trans_time;
 
     uint8_t _comms_received[CONST_MAX_SIZE_BUFFER];
     uint8_t _comms_received_size;
